+++ src/js/content/audio/scan.js
@@ -57,9 +57,8 @@ content.audio.scan = (() => {
   }
 
   function renderGroup(group = [], {pan, when} = {}) {
-    const duration = 2
-
-    const panner = context.createStereoPanner()
+    const duration = 2,
+      panner = context.createStereoPanner()
 
     panner.pan.value = pan
     panner.connect(lowpass)
@@ -73,11 +72,12 @@ content.audio.scan = (() => {
 
     for (let i = 0; i < count; i += 1) {
       const gain = (1 - (i / (count - 1))) ** 4,
-        next = when + ((i + 1) * (duration / count)),
-        note = engine.utility.scale(group[i], 0, 100/12, 0, 12)
+        next = when + ((i + 1) * (duration / count))
 
-      const detune = (note - Math.round(note)) * 100,
-        frequency = content.utility.frequency.fromMidi(engine.utility.clamp(60 + Math.round(note), 0, 127))
+      const {
+        detune,
+        frequency,
+      } = toNote(group[i])
 
       synth.param.detune.linearRampToValueAtTime(detune, next)
       synth.param.frequency.exponentialRampToValueAtTime(frequency, next)
@@ -87,9 +87,80 @@ content.audio.scan = (() => {
     synth.stop(when + duration)
   }
 
+  function renderNearbyMaterial(relative) {
+    const distance = engine.utility.clamp(relative.subtract({z: relative.z}).distance() / 100, 0, 1)
+
+    const color = engine.utility.lerpExp(1, 8, Math.max(0, Math.cos(Math.atan2(relative.y, relative.x))), 0.5),
+      delay = distance * 2,
+      duration = 1/12,
+      gain = engine.utility.fromDb(engine.utility.lerp(-3, -12, distance)),
+      when = engine.audio.time() + delay
+
+    const {
+      detune,
+      frequency,
+    } = toNote(relative.z)
+
+    const synth = engine.audio.synth.createSimple({
+      detune,
+      frequency,
+      type: 'square',
+      when,
+    }).filtered({
+      detune,
+      frequency: frequency * color,
+    })
+
+    const binaural = engine.audio.binaural.create({
+      ...relative.normalize(),
+    }).from(synth).to(bus)
+
+    synth.param.gain.setValueAtTime(engine.const.zeroGain, when)
+    synth.param.gain.exponentialRampToValueAtTime(gain, when + 1/32)
+    synth.param.gain.linearRampToValueAtTime(engine.const.zeroGain, when + duration)
+    synth.stop(when + duration)
+
+    setTimeout(() => binaural.destroy(), (delay + duration) * 1000)
+  }
+
+  function renderNearbyMaterials() {
+    const position = engine.position.getVector(),
+      quaternion = engine.position.getQuaternion().conjugate(),
+      renderDistance = 100
+
+    const materials = content.materials.nearby.retrieveAll({
+      depth: renderDistance * 2,
+      height: renderDistance * 2,
+      width: renderDistance * 2,
+      x: position.x - renderDistance,
+      y: position.y - renderDistance,
+      z: position.z - renderDistance,
+    })
+
+    for (const material of materials) {
+      const relative = engine.utility.vector3d.create(material)
+        .subtract(position)
+        .rotateQuaternion(quaternion)
+
+      renderNearbyMaterial(relative)
+    }
+  }
+
+  function toNote(z = 0) {
+    const scale = 100
+
+    const note = engine.utility.scale(z, 0, scale/12, 0, 12)
+
+    return {
+      detune: (note - Math.round(note)) * scale,
+      frequency: content.utility.frequency.fromMidi(engine.utility.clamp(60 + Math.round(note), 0, 127)),
+    }
+  }
+
   return {
     render: function (scan) {
       render(scan)
+      renderNearbyMaterials()
       return this
     },
   }